# modules/swagger-jaxrs-client-kek/swagger-inputs/swagger.yaml
swagger: "2.0"
info:
  description: "BaaS for delivery services like Glovo, Uber or even for regular mail. Also, a simple web front-end should be provided as an example of a typical consumer's app."
  version: "1.0.0"
  title: "Kinda Express King"
#  contact:
#    email: "apiteam@softserveinc.com"
# host: "localhost:8080"
...
paths:
...
  /tenants:
    get:
      tags:
        - "tenants"
      summary: "Searches for tenants"
      description: "Finds all tenants"
      operationId: "getTenantList"
      produces:
        - "application/vnd.softserve.tenantList+json"
      responses:
        200:
          description: "List of tenants"
          schema:
            $ref: "#/definitions/tenantList"
    post:
      tags:
        - "tenants"
      summary: "Creates a new tenant"
      operationId: "addTenant"
      consumes:
        - "application/vnd.softserve.tenant+json"
      produces:
        - "application/vnd.softserve.tenant+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "tenant"
          in: "body"
          description: "The tenant to create"
          schema:
            $ref: "#/definitions/tenant"
      responses:
        200:
          description: "The newly created tenant object"
          schema:
            $ref: "#/definitions/tenant"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
  /tenants/{guid}:
    get:
      tags:
        - "tenants"
      summary: "Finds the specific tenant"
      operationId: "getTenant"
      produces:
        - "application/vnd.softserve.tenant+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
      responses:
        200:
          description: "The tenant object"
          schema:
            $ref: "#/definitions/tenant"
    put:
      tags:
        - "tenants"
      summary: "Modifies the specific tenant"
      operationId: "modifyTenant"
      consumes:
        - "application/vnd.softserve.tenant+json"
      produces:
        - "application/vnd.softserve.tenant+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "tenant"
          in: "body"
          description: "The tenant to modify"
          schema:
            $ref: "#/definitions/tenant"
      responses:
        200:
          description: "The modified tenant object"
          schema:
            $ref: "#/definitions/tenant"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
    delete:
      tags:
        - "tenants"
      summary: "Deletes the specific tenant"
      operationId: "deleteTenant"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
  /tenants/{guid}/properties:
    get:
      tags:
        - "tenants"
      summary: "Finds properties of the specific tenant"
      operationId: "getTenantProperties"
      produces:
        - "application/vnd.softserve.tenantPropertyList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
      responses:
        200:
          description: "List of tenant properties"
          schema:
            $ref: "#/definitions/tenantPropertyList"
    post:
      tags:
        - "tenants"
      summary: "Adds tenant properties"
      operationId: "addTenantProperties"
      consumes:
        - "application/vnd.softserve.tenantPropertyList+json"
      produces:
        - "application/vnd.softserve.tenantPropertyList+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "properties"
          in: "body"
          description: "List of properties"
          schema:
            $ref: "#/definitions/tenantPropertyList"
      responses:
        200:
          description: "List of added tenant properties"
          schema:
            $ref: "#/definitions/tenantPropertyList"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
  /tenants/{guid}/properties/{propguid}:
    get:
      tags:
        - "tenants"
      summary: "Finds specific property of the specific tenant"
      operationId: "getTenantProperty"
      produces:
        - "application/vnd.softserve.tenantProperty+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "propguid"
          in: "path"
          description: "ID of the specific tenant property"
          required: true
          type: "string"
      responses:
        200:
          description: "Specific tenant tenant property"
          schema:
            $ref: "#/definitions/tenantProperty"
    put:
      tags:
        - "tenants"
      summary: "Modifies the specific tenant property"
      operationId: "modifyTenantProperty"
      consumes:
        - "application/vnd.softserve.tenantProperty+json"
      produces:
        - "application/vnd.softserve.tenantProperty+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "propguid"
          in: "path"
          description: "ID of the specific tenant property"
          required: true
          type: "string"
        - name: "property"
          in: "body"
          description: "The tenant property to modify"
          schema:
            $ref: "#/definitions/tenantProperty"
      responses:
        200:
          description: "The modified tenant property object"
          schema:
            $ref: "#/definitions/tenantProperty"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
    delete:
      tags:
        - "tenants"
      summary: "Deletes the specific tenant property"
      operationId: "deleteTenantProperty"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "propguid"
          in: "path"
          description: "ID of the specific tenant property"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
  /tenants/{guid}/addresses:
    get:
      tags:
        - "tenants"
      summary: "Finds addressess of the specific tenant"
      operationId: "getTenantAddresses"
      produces:
        - "application/vnd.softserve.addressList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
      responses:
        200:
          description: "List of tenant addresses"
          schema:
            $ref: "#/definitions/addressList"
    post:
      tags:
        - "tenants"
      summary: "Adds a new addresses"
      operationId: "addTenantAddresses"
      consumes:
        - "application/vnd.softserve.addressList+json"
      produces:
        - "application/vnd.softserve.addressList+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "addresses"
          in: "body"
          description: "List of addresses"
          schema:
            $ref: "#/definitions/addressList"
      responses:
        200:
          description: "List of added tenant addresses"
          schema:
            $ref: "#/definitions/addressList"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
  /tenants/{guid}/addresses/{addrguid}:
    get:
      tags:
        - "tenants"
      summary: "Finds specific address of the specific tenant"
      operationId: "getTenantAddress"
      produces:
        - "application/vnd.softserve.address+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "addrguid"
          in: "path"
          description: "ID of the specific address"
          required: true
          type: "string"
      responses:
        200:
          description: "Specific tenant address"
          schema:
            $ref: "#/definitions/address"
    put:
      tags:
        - "tenants"
      summary: "Modifies the specific tenant address"
      operationId: "modifyTenantAddress"
      consumes:
        - "application/vnd.softserve.address+json"
      produces:
        - "application/vnd.softserve.address+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "addrguid"
          in: "path"
          description: "ID of the specific address"
          required: true
          type: "string"
        - name: "address"
          in: "body"
          description: "The address to modify"
          schema:
            $ref: "#/definitions/address"
      responses:
        200:
          description: "The modified address object"
          schema:
            $ref: "#/definitions/address"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
    delete:
      tags:
        - "tenants"
      summary: "Deletes the specific tenant address"
      operationId: "deleteTenantAddress"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific tenant"
          required: true
          type: "string"
        - name: "addrguid"
          in: "path"
          description: "ID of the specific address"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
...
definitions:
...
  tenantList:
    type: "object"
    properties:
      tenantList:
        type: "array"
        items:
          $ref: "#/definitions/tenant"
  tenant:
    type: "object"
    properties:
      guid:
        type: "string"
      owner:
        type: "string"
        description: "tenant owner ID"
      name:
        type: "string"
      details:
        $ref: "#/definitions/tenantDetails"
  tenantDetails:
    type: "object"
    properties:
      payload:
        type: "string"
      imageURL:
        type: "string"
  tenantPropertyList:
    type: "object"
    properties:
      tenantPropertyList:
        type: "array"
        items:
          $ref: "#/definitions/tenantProperty"
  tenantProperty:
    type: "object"
    properties:
      guid:
        type: "string"
      tenant:
        type: "string"
      type:
        type: "string"
      key:
        type: "string"
      value:
        type: "string"
...
  addressList:
    type: "object"
    properties:
      addressList:
        type: "array"
        items:
          $ref: "#/definitions/address"
  address:
    type: "object"
    properties:
      guid:
        type: "string"
      alias:
        type: "string"
      address:
        type: "string"
      notes:
        type: "string"
  errorList:
    type: "object"
    properties:
      errorlList:
        type: "array"
        items:
          $ref: "#/definitions/error"
  error:
    type: "object"
    properties:
      error:
        type: "string"
      details:
        type: "string"

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /burnkr/kek/1.0.0
schemes:
  - https